Webdriver tests
Add Login and Pool tests to all_tests
Add screenshot dir to config files
Catch exceptions and print stack trace and save screenshot with a
timestamp and test name in the filename.
