+++ README.md
@@ -1,17 +1,16 @@
 
 Check out the code
-===================
+    git@github.com:rockstor/rockstor-tests.git
 
 Install required modules
-==========================
     pip install -r requirements.txt
 
 Configure
-======================================================
+
 Set the rockstor base_url in webdriver/development.yaml
 
 Run tests
-=========
+
 cd to the webdriver directory
 
 You can run an individual test like this
