--- webdriver/config.yaml
@@ -1,1 +1,0 @@
-base_url: https://rockstor-vm
