+++ webdriver/development.yaml
@@ -1,2 +1,2 @@
-base_url: https://rockstor-iso.mydomain.com
-
+base_url: https://rs14dev
+screenshot_dir: output
