+++ webdriver/staging.yaml
@@ -1,2 +1,3 @@
-base_url: https://rockstor-vm
+base_url: https://rs14staging
+screenshot_dir: output
